Move payment method components inside the account settings folder
